+++ src/controllers/mainController.ts
@@ -10,7 +10,7 @@ import * as vscode from 'vscode';
 import { IConnectionInfo } from 'vscode-mssql';
 import { AzureResourceController } from '../azure/azureResourceController';
 import * as Constants from '../constants/constants';
-import * as LocalizedConstants from '../constants/localizedConstants';
+import * as LocalizedConstants from '../constants/locConstants';
 import SqlToolsServerClient from '../languageservice/serviceclient';
 import * as ConnInfo from '../models/connectionInfo';
 import { CompletionExtensionParams, CompletionExtLoadRequest, RebuildIntelliSenseNotification } from '../models/contracts/languageService';
@@ -42,9 +42,13 @@ import VscodeWrapper from './vscodeWrapper';
 import { sendActionEvent } from '../telemetry/telemetry';
 import { TelemetryActions, TelemetryViews } from '../telemetry/telemetryInterfaces';
 import { TableDesignerService } from '../services/tableDesignerService';
-import { TableDesignerWebViewController } from '../tableDesigner/tableDesignerWebViewController';
-import { ConnectionDialogWebViewController } from '../connectionconfig/connectionDialogWebViewController';
+import { TableDesignerWebviewController } from '../tableDesigner/tableDesignerWebviewController';
+import { ConnectionDialogWebviewController } from '../connectionconfig/connectionDialogWebviewController';
+import { ObjectExplorerFilter } from '../objectExplorer/objectExplorerFilter';
+import { ExecutionPlanService } from '../services/executionPlanService';
+import { ExecutionPlanWebviewController } from './executionPlanWebviewController';
 import { QueryResultWebViewController } from '../queryResult/queryResultWebViewController';
+
 
 /**
  * The main controller class that initializes the extension
@@ -77,6 +81,7 @@ export default class MainController implements vscode.Disposable {
 	public tableDesignerService: TableDesignerService;
 	public configuration: vscode.WorkspaceConfiguration;
 	public objectExplorerTree: vscode.TreeView<TreeNodeInfo>;
+	public executionPlanService: ExecutionPlanService;
 
 	/**
 	 * The main controller constructor
@@ -91,7 +96,7 @@ export default class MainController implements vscode.Disposable {
 		}
 		this._vscodeWrapper = vscodeWrapper || new VscodeWrapper();
 		this._untitledSqlDocumentService = new UntitledSqlDocumentService(this._vscodeWrapper);
-		this.configuration = vscode.workspace.getConfiguration(Constants.extensionName);
+		this.configuration = vscode.workspace.getConfiguration();
 	}
 
 	/**
@@ -208,6 +213,10 @@ export default class MainController implements vscode.Disposable {
 			this.azureAccountService = new AzureAccountService(this._connectionMgr.azureController, this._connectionMgr.accountStore);
 			this.azureResourceService = new AzureResourceService(this._connectionMgr.azureController, azureResourceController, this._connectionMgr.accountStore);
 			this.tableDesignerService = new TableDesignerService(SqlToolsServerClient.instance);
+			this.executionPlanService = new ExecutionPlanService(SqlToolsServerClient.instance);
+
+			const providerInstance = new this.ExecutionPlanCustomEditorProvider(this._context, this.executionPlanService, this._untitledSqlDocumentService);
+			vscode.window.registerCustomEditorProvider('mssql.executionPlanView', providerInstance);
 
 			// Add handlers for VS Code generated commands
 			this._vscodeWrapper.onDidCloseTextDocument(async (params) => await this.onDidCloseTextDocument(params));
@@ -450,7 +459,7 @@ export default class MainController implements vscode.Disposable {
 				}
 				await self.createObjectExplorerSession();
 			} else {
-				const connDialog = new ConnectionDialogWebViewController(
+				const connDialog = new ConnectionDialogWebviewController(
 					this._context,
 					this,
 					this._objectExplorerProvider
@@ -526,7 +535,7 @@ export default class MainController implements vscode.Disposable {
 			this._context.subscriptions.push(
 				vscode.commands.registerCommand(
 					Constants.cmdEditConnection, async (node: TreeNodeInfo) => {
-						const connDialog = new ConnectionDialogWebViewController(
+						const connDialog = new ConnectionDialogWebviewController(
 							this._context,
 							this,
 							this._objectExplorerProvider,
@@ -539,7 +548,7 @@ export default class MainController implements vscode.Disposable {
 			this._context.subscriptions.push(
 				vscode.commands.registerCommand(
 					Constants.cmdNewTable, async (node: TreeNodeInfo) => {
-						const reactPanel = new TableDesignerWebViewController(
+						const reactPanel = new TableDesignerWebviewController(
 							this._context,
 							this.tableDesignerService,
 							this._connectionMgr,
@@ -552,7 +561,7 @@ export default class MainController implements vscode.Disposable {
 			this._context.subscriptions.push(
 				vscode.commands.registerCommand(
 					Constants.cmdEditTable, async (node: TreeNodeInfo) => {
-						const reactPanel = new TableDesignerWebViewController(
+						const reactPanel = new TableDesignerWebviewController(
 							this._context,
 							this.tableDesignerService,
 							this._connectionMgr,
@@ -564,6 +573,41 @@ export default class MainController implements vscode.Disposable {
 
 			this._context.subscriptions.push(
 				vscode.window.registerWebviewViewProvider("queryResult", this._queryResultWebviewController));
+
+			const filterNode = async (node: TreeNodeInfo) => {
+				const filters = await ObjectExplorerFilter.getFilters(this._context, node);
+				if (filters) {
+					node.filters = filters;
+					if(node.collapsibleState === vscode.TreeItemCollapsibleState.Collapsed) {
+						await this._objectExplorerProvider.refreshNode(node);
+					} else if (node.collapsibleState === vscode.TreeItemCollapsibleState.Expanded) {
+						await this._objectExplorerProvider.expandNode(node, node.sessionId, undefined);
+					}
+					await this.objectExplorerTree.reveal(node, {select: true, focus: true, expand: true});
+				} else {
+					// User cancelled the operation. Do nothing and focus on the node
+					await this.objectExplorerTree.reveal(node, {select: true, focus: true});
+					return;
+				}
+			};
+
+			this._context.subscriptions.push(
+				vscode.commands.registerCommand(
+					Constants.cmdFilterNode, filterNode));
+
+			this._context.subscriptions.push(
+				vscode.commands.registerCommand(
+					Constants.cmdFilterNodeWithExistingFilters, filterNode));
+
+			this._context.subscriptions.push(
+				vscode.commands.registerCommand(
+					Constants.cmdClearFilters, async(node: TreeNodeInfo) => {
+						node.filters = [];
+						await this._objectExplorerProvider.refreshNode(node);
+						await this.objectExplorerTree.reveal(node, {select: true, focus: true, expand: true});
+					})
+				);
+
 		}
 
 		// Initiate the scripting service
@@ -605,10 +649,10 @@ export default class MainController implements vscode.Disposable {
 			vscode.commands.registerCommand(Constants.cmdCopyObjectName, async () => {
 				let node = this._objectExplorerProvider.currentNode;
 				// Folder node
-				if (node.contextValue === Constants.folderLabel) {
+				if (node.context.type === Constants.folderLabel) {
 					return;
-				} else if (node.contextValue === Constants.serverLabel ||
-					node.contextValue === Constants.disconnectedServerLabel) {
+				} else if (node.context.type === Constants.serverLabel ||
+					node.context.type === Constants.disconnectedServerLabel) {
 					const label = typeof node.label === 'string' ? node.label : node.label.label;
 					await this._vscodeWrapper.clipboardWriteText(label);
 				} else {
@@ -1258,6 +1302,10 @@ export default class MainController implements vscode.Disposable {
 			this._statusview.languageFlavorChanged(doc.uri.toString(true), Constants.mssqlProviderName);
 		}
 
+		if (doc && doc.languageId === Constants.sqlPlanLanguageId) {
+			vscode.commands.executeCommand('workbench.action.closeActiveEditor');
+		}
+
 		// Setup properties incase of rename
 		this._lastOpenedTimer = new Utils.Timer();
 		this._lastOpenedTimer.start();
@@ -1383,6 +1431,10 @@ export default class MainController implements vscode.Disposable {
 			if (e.affectsConfiguration(Constants.enableConnectionPooling)) {
 				await this.displayReloadMessage(LocalizedConstants.reloadPromptGeneric);
 			}
+
+			if(e.affectsConfiguration(Constants.configEnableExperimentalFeatures)) {
+				await this.displayReloadMessage(LocalizedConstants.reloadPromptGeneric);
+			}
 		}
 	}
 
@@ -1420,4 +1472,41 @@ export default class MainController implements vscode.Disposable {
 	public onClearAzureTokenCache(): void {
 		this.connectionManager.onClearTokenCache();
 	}
+
+	private ExecutionPlanCustomEditorProvider = class implements vscode.CustomTextEditorProvider {
+		context: vscode.ExtensionContext;
+		executionPlanService: ExecutionPlanService;
+		untitledSqlService: UntitledSqlDocumentService
+        constructor(
+			context: vscode.ExtensionContext,
+			executionPlanService: ExecutionPlanService,
+			untitledSqlService: UntitledSqlDocumentService
+        ) {
+			this.context = context;
+			this.executionPlanService = executionPlanService;
+			this.untitledSqlService = untitledSqlService;
+        }
+
+        public async resolveCustomTextEditor(
+            document: vscode.TextDocument
+        ): Promise<void> {
+			await this.onOpenExecutionPlanFile(document);
+        }
+
+		public async onOpenExecutionPlanFile(document: vscode.TextDocument) {
+			const planContents = document.getText();
+			let docName = document.fileName;
+			docName = docName.substring(docName.lastIndexOf(path.sep) + 1);
+
+			const executionPlanController = new ExecutionPlanWebviewController(
+				this.context,
+				this.executionPlanService,
+				this.untitledSqlService,
+				planContents,
+				docName
+			)
+
+			executionPlanController.revealToForeground();
+		}
+    };
 }