+++ src/objectExplorer/connectTreeNode.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-import * as LocalizedConstants from '../constants/localizedConstants';
+import * as LocalizedConstants from '../constants/locConstants';
 import * as Constants from '../constants/constants';
 import { TreeNodeInfo } from './treeNodeInfo';