+++ src/objectExplorer/treeNodeInfo.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
+import * as vscodeMssql from 'vscode-mssql';
 import { NodeInfo } from '../models/contracts/objectExplorer/nodeInfo';
 import { ObjectExplorerUtils } from './objectExplorerUtils';
 import * as Constants from '../constants/constants';
@@ -21,10 +22,13 @@ export class TreeNodeInfo extends vscode.TreeItem implements ITreeNodeInfo {
 	private _parentNode: TreeNodeInfo;
 	private _connectionInfo: IConnectionInfo;
 	private _metadata: ObjectMetadata;
+	private _filterableProperties: vscodeMssql.NodeFilterProperty[];
+	private _filters: vscodeMssql.NodeFilter[];
+	private _originalLabel: string;
 
 	constructor(
 		label: string,
-		contextValue: string,
+		context: vscodeMssql.TreeNodeContextValue,
 		collapsibleState: vscode.TreeItemCollapsibleState,
 		nodePath: string,
 		nodeStatus: string,
@@ -32,17 +36,22 @@ export class TreeNodeInfo extends vscode.TreeItem implements ITreeNodeInfo {
 		sessionId: string,
 		connectionInfo: IConnectionInfo,
 		parentNode: TreeNodeInfo,
-		objectMetadata?: ObjectMetadata
+		filterProperties: vscodeMssql.NodeFilterProperty[],
+		objectMetadata?: ObjectMetadata,
+		filters?: vscodeMssql.NodeFilter[]
 	) {
 		super(label, collapsibleState);
-		this.contextValue = contextValue;
+		this._originalLabel = label;
+		this.context = context;
 		this._nodePath = nodePath;
 		this._nodeStatus = nodeStatus;
 		this._nodeType = nodeType;
 		this._sessionId = sessionId;
 		this._parentNode = parentNode;
 		this._connectionInfo = connectionInfo;
+		this._filterableProperties = filterProperties;
 		this._metadata = objectMetadata;
+		this._filters = filters;
 		this.iconPath = ObjectExplorerUtils.iconPath(this.nodeType);
 	}
 
@@ -54,16 +63,26 @@ export class TreeNodeInfo extends vscode.TreeItem implements ITreeNodeInfo {
 		label?: string,
 		nodeType?: string): TreeNodeInfo {
 		let type = nodeType ? nodeType : nodeInfo.nodeType;
-		if((nodeInfo as any).objectType === "Tables") {
-			type = 'TablesFolder';
-		}
-		const treeNodeInfo = new TreeNodeInfo(label ? label : nodeInfo.label, type,
-			nodeInfo.isLeaf ? vscode.TreeItemCollapsibleState.None :
-				(type === Constants.serverLabel ? vscode.TreeItemCollapsibleState.Expanded :
-					vscode.TreeItemCollapsibleState.Collapsed),
-			nodeInfo.nodePath, nodeInfo.nodeStatus,
-			type, sessionId, connectionInfo, parentNode, nodeInfo.metadata);
-		console.log('TreeNode', treeNodeInfo, nodeInfo);
+
+
+		const treeNodeInfo = new TreeNodeInfo(
+			label ? label : nodeInfo.label,
+			{
+				type: type,
+				filterable: nodeInfo.filterableProperties?.length > 0,
+				hasFilters: false,
+				subType: nodeInfo.objectType
+			},
+			nodeInfo.isLeaf ? vscode.TreeItemCollapsibleState.None : (type === Constants.serverLabel ? vscode.TreeItemCollapsibleState.Expanded : vscode.TreeItemCollapsibleState.Collapsed),
+			nodeInfo.nodePath,
+			nodeInfo.nodeStatus,
+			type,
+			sessionId,
+			connectionInfo,
+			parentNode,
+			nodeInfo.filterableProperties,
+			nodeInfo.metadata
+		);
 		return treeNodeInfo;
 	}
 
@@ -108,6 +127,18 @@ export class TreeNodeInfo extends vscode.TreeItem implements ITreeNodeInfo {
 		return this._metadata;
 	}
 
+	public get filterableProperties(): vscodeMssql.NodeFilterProperty[] {
+		return this._filterableProperties;
+	}
+
+	public get context(): vscodeMssql.TreeNodeContextValue {
+		return this._convertToTreeNodeContext(this.contextValue);
+	}
+
+	public get filters(): vscodeMssql.NodeFilter[] {
+		return this._filters;
+	}
+
 	/** Setters */
 	public set nodePath(value: string) {
 		this._nodePath = value;
@@ -119,6 +150,7 @@ export class TreeNodeInfo extends vscode.TreeItem implements ITreeNodeInfo {
 
 	public set nodeType(value: string) {
 		this._nodeType = value;
+		this._updateContextValue();
 	}
 
 	public set nodeSubType(value: string) {
@@ -143,5 +175,55 @@ export class TreeNodeInfo extends vscode.TreeItem implements ITreeNodeInfo {
 
 	public set connectionInfo(value: IConnectionInfo) {
 		this._connectionInfo = value;
+	}
+
+	public set filterableProperties(value: vscodeMssql.NodeFilterProperty[]) {
+		this._filterableProperties = value;
+		this._updateContextValue();
+	}
+
+	public set filters(value: vscodeMssql.NodeFilter[]) {
+		this._filters = value;
+		this._updateContextValue();
+		this.label = value.length > 0 ? vscode.l10n.t('{0} (filtered)', this._originalLabel) : this._originalLabel;
+	}
+
+	public set context(value: vscodeMssql.TreeNodeContextValue)  {
+		this.contextValue = this._convertToContextValue(value);
+	}
+
+	private _updateContextValue() {
+		const contextValue = this.context;
+		contextValue.filterable = this.filterableProperties?.length > 0;
+		contextValue.hasFilters = this.filters?.length > 0;
+		this.context = contextValue;
+	}
+
+	//split the context value with, and is in the form of key=value and convert it to TreeNodeContextValue
+	private _convertToTreeNodeContext(contextValue: string): vscodeMssql.TreeNodeContextValue {
+		let contextArray = contextValue.split(',');
+		let context: vscodeMssql.TreeNodeContextValue = {
+			filterable: false,
+			hasFilters: false,
+			type: undefined,
+			subType: undefined
+		};
+		contextArray.forEach(element => {
+			let keyValuePair = element.split('=');
+			context[keyValuePair[0]] = keyValuePair[1];
+		});
+		return context;
+	}
+
+	//convert TreeNodeContextValue to context value string
+	private _convertToContextValue(context: vscodeMssql.TreeNodeContextValue): string {
+		if (context === undefined) {
+			return '';
+		}
+		let contextValue = '';
+		Object.keys(context).forEach(key => {
+			contextValue += key + '=' + context[key] + ',';
+		});
+		return contextValue;
 	}
 }