+++ test/unit/perFileConnection.test.ts
@@ -12,7 +12,7 @@ import SqlToolsServiceClient from '../../src/languageservice/serviceclient';
 
 import { IConnectionInfo, IServerInfo } from 'vscode-mssql';
 import * as Constants from '../../src/constants/constants';
-import * as LocalizedConstants from '../../src/constants/localizedConstants';
+import * as LocalizedConstants from '../../src/constants/locConstants';
 import ConnectionManager from '../../src/controllers/connectionManager';
 import MainController from '../../src/controllers/mainController';
 import VscodeWrapper from '../../src/controllers/vscodeWrapper';
@@ -59,6 +59,9 @@ function createTestCredentials(): IConnectionInfo {
 		trustServerCertificate: false,
 		hostNameInCertificate: '',
 		persistSecurityInfo: false,
+		columnEncryptionSetting: 'enabled',
+		attestationProtocol: 'HGS',
+		enclaveAttestationUrl: 'https://attestationurl',
 		connectTimeout: 15,
 		commandTimeout: 30,
 		connectRetryCount: 0,